iprt/armv8.h: Build fix. bugref:10371
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@101544 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ include/iprt/armv8.h
@@ -3007,7 +3007,7 @@ typedef enum ARMV8INSTRCOND
  */
 DECL_FORCE_INLINE(uint32_t) Armv8A64MkInstrBCond(ARMV8INSTRCOND enmCond, int32_t iImm19)
 {
-    Assert(enmCond >= 0 && enmCond < 16);
+    Assert((int)enmCond >= 0 && (int)enmCond < 16);
     return UINT32_C(0x54000000)
          | (((uint32_t)iImm19 & 0x7ffff) <<  5)
          | (uint32_t)enmCond;
